refactor: extract renderQuestion to reusable utility
- Create src/components/survey-editor/utils/questionRenderer.tsx
- Remove duplicate renderQuestion code from ParallelGroupRenderer.tsx (357  200 lines)
- Remove duplicate renderQuestion code from PagePreview.tsx
- Fix: maintain exact same functionality and UI
- Fix: preserve all question types support (Text, Radio, Checkbox, etc.)
- Improve: DRY principle - one place for question rendering logic

Breaking changes: NONE
Visual changes: NONE
