+++ src/components/survey-preview/PagePreview.tsx
@@ -1,16 +1,9 @@
-import React, { useState } from 'react';
-import { Question, QUESTION_TYPES, ParallelBranchSettings, TransitionRule } from '@survey-platform/shared-types';
-import { Input } from '@/components/ui/input';
+import React from 'react';
+import { Question, QUESTION_TYPES } from '@survey-platform/shared-types';
 import { Label } from '@/components/ui/label';
-import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group';
-import { Checkbox } from '@/components/ui/checkbox';
-import { Calendar } from '@/components/ui/calendar';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
-import { IMaskInput } from 'react-imask';
-import { format } from 'date-fns';
-import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import { PlaceholderText } from '@/components/ui/placeholder-text';
 import { ParallelGroupRenderer } from '@/components/survey-editor/ParallelGroupRenderer';
+import { renderQuestion } from '@/components/survey-editor/utils/questionRenderer';
 
 interface PagePreviewProps {
   questions: Question[];
@@ -172,150 +165,4 @@ export function PagePreview({ questions, answers, onAnswerChange, pages, pageId,
       </div>
     </div>
   );
-}
-
-// Функция для рендеринга обычных вопросов (не параллельных групп)
-function renderQuestion(
-  question: Question,
-  questions: Question[],
-  answers: Record<string, any>,
-  answerHandler: (id: string, value: any) => void
-) {
-  // Получаем ответ для текущего вопроса
-  const currentAnswer = answers[question.id];
-
-  switch (question.type) {
-    case QUESTION_TYPES.ParallelGroup: {
-      // Этот случай не должен вызываться, так как ParallelGroup обрабатывается отдельно
-      console.warn('[renderQuestion] ParallelGroup должен обрабатываться в ParallelGroupRenderer');
-      return null;
-    }
-
-    case QUESTION_TYPES.Text:
-      return (
-        <Input
-          value={currentAnswer || ''}
-          onChange={(e) => answerHandler(question.id, e.target.value)}
-          placeholder="Введите ваш ответ"
-        />
-      );
-
-    case QUESTION_TYPES.Number:
-      return (
-        <Input
-          type="number"
-          value={currentAnswer || ''}
-          onChange={(e) => answerHandler(question.id, e.target.value)}
-          placeholder="Введите число"
-        />
-      );
-
-    case QUESTION_TYPES.Radio:
-      return (
-        <RadioGroup
-          value={currentAnswer || ''}
-          onValueChange={(value) => answerHandler(question.id, value)}
-        >
-          {question.options?.map((option) => (
-            <div key={option.id} className="flex items-center space-x-2">
-              <RadioGroupItem value={option.id} id={option.id} />
-              <Label htmlFor={option.id}>{option.text}</Label>
-            </div>
-          ))}
-        </RadioGroup>
-      );
-
-    case QUESTION_TYPES.Checkbox:
-      return (
-        <div className="space-y-2">
-          {question.options?.map((option) => (
-            <div key={option.id} className="flex items-center space-x-2">
-              <Checkbox
-                id={option.id}
-                checked={currentAnswer?.includes(option.id) || false}
-                onCheckedChange={(checked) => {
-                  const currentAnswers = currentAnswer || [];
-                  answerHandler(
-                    question.id,
-                    checked
-                      ? [...currentAnswers, option.id]
-                      : currentAnswers.filter((id: string) => id !== option.id)
-                  );
-                }}
-              />
-              <Label htmlFor={option.id}>{option.text}</Label>
-            </div>
-          ))}
-        </div>
-      );
-
-    case QUESTION_TYPES.Select:
-      return (
-        <Select
-          value={currentAnswer || ''}
-          onValueChange={(value) => answerHandler(question.id, value)}
-        >
-          <SelectTrigger>
-            <SelectValue placeholder="Выберите вариант" />
-          </SelectTrigger>
-          <SelectContent>
-            {question.options?.map((option) => (
-              <SelectItem key={option.id} value={option.id}>
-                {option.text}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-      );
-
-    case QUESTION_TYPES.Date:
-      const dateSettings = question.settings as { format?: string };
-      return (
-        <div className="space-y-2">
-          <Calendar
-            mode="single"
-            selected={currentAnswer}
-            onSelect={(value) => answerHandler(question.id, value)}
-            className="rounded-md border"
-          />
-          {currentAnswer && (
-            <p className="text-sm text-gray-500">
-              Выбрано: {format(currentAnswer, dateSettings?.format || 'dd.MM.yyyy')}
-            </p>
-          )}
-        </div>
-      );
-
-    case QUESTION_TYPES.Email:
-      return (
-        <Input
-          type="email"
-          value={currentAnswer || ''}
-          onChange={(e) => answerHandler(question.id, e.target.value)}
-          placeholder="example@domain.com"
-        />
-      );
-
-    case QUESTION_TYPES.Phone:
-      const phoneSettings = question.settings as { countryCode?: string; mask?: string };
-      return (
-        <div className="flex gap-2">
-          <Input
-            className="w-20"
-            value={phoneSettings?.countryCode || '+7'}
-            disabled
-          />
-          <IMaskInput
-            className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
-            mask={phoneSettings?.mask || '(000) 000-00-00'}
-            value={currentAnswer || ''}
-            onAccept={(value) => answerHandler(question.id, value)}
-            placeholder="(999) 999-99-99"
-          />
-        </div>
-      );
-
-    default:
-      return null;
-  }
 }
